Adjust phone layouts for phablet and tablet UIs
This change adds an extra top padding to the FragmentActionBarPadding style in layouts equeals or bigger than 580dp.

This solution is the simplest one and it still has a little glitch.
I've used the attribute actionBarSize to add the extra padding but in 680dp it seems that it's not equal to the actual action bar height so there's a
little black space

The correct solution should be rework all the app layout (the Phone part) to adapt better to bigger layouts.
This solution will be never implemented by google because, obviously, Phone.apk is not meant to work at layouts different than 360dp (phoneui)

Change-Id: I1e162f5f9d7b51e1892ad93357fb74451fa1de26